requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
coveralls==2.0.0
distlib==0.3.0
Django==3.0.5
django-crispy-forms==1.9.0
django-rest-marshmallow==4.0.2
djangorestframework==3.11.0
docopt==0.6.2
drf-yasg==1.17.1
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
flake8-docstrings==1.5.0
flake8-import-order==0.18.1
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
identify==1.4.14
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
inflection==0.4.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
marshmallow==3.5.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.2.0
nltk==3.5
nodeenv==1.3.5
openapi-codec==1.3.2
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.2.0
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
pyparsing==2.4.7
python-coveralls==2.9.3
python-dotenv==0.13.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
rope==0.16.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.0
tornado==6.0.4
tqdm==4.45.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.18
wcwidth==0.1.9
Werkzeug==1.0.1
whitenoise==5.0.1
zipp==3.1.0
